Can read input string and tokenize into individual expressions. (see body)
Will need to do more specific checking and leave as much info
as possible so that the parser can analyze the expressions.

So far, we iterate through the chars of the input string,
and if the first character or the last character is a '"',
it is skipped (will not be part of the expression string)

Otherwise, once a ';' is hit, a duplicate of the string is made
to live on the heap, and a null-terminator is placed where
the ';' was (on the duplicate string). Then, the pointer
assigned to the heap-allocated string duplicate is
assigned to svec, a (char **), which is a vector of (char *)

If svec hits its max capacity, the memory is realloc'ed
by doubling its space.

Once all the expressions make it into svec,
they are ready to be analyzed by the parser, one string (expression)
at a time.