No need for first_char boolean during vector_str population (see body)
If the first char of the input string is '"',
just advance pos one byte over so that the '"' char is not included
with the first expression tokenized.

Likewise, if the last char of the input string is '"'
(prior to the null-terminator),
add a null terminator where that '"' resides so that character
is omitted from tokenization.