Added macros is_header_free and is_header_used to prepare for making size field a type int16_t (signed short) -- the size field, in addition to storing the size of the block proceeding it, will denote whether the block is in use by the client, or is free -- by its sign.
Let curr be a variable of type header_t.
If curr->size == -128, the proceeding block is 128 bytes, and it is in use.
Likewise, if curr->size == 128, the proceeding block is 128 bytes, and it is free.
Therefore, a size value less than 0 denotes a proceeding block that is in use,
and a size value greater than 0 denotes a proceeding block that is free.
Sizes of 0 are invalid, and will not be allowed by mymalloc.

This is the idea in mind that will reduce header_t from
4 bytes, down to 2 bytes. (the only field will be header_t size).