Fixed randrnge macro, it wasn't [min, max) like I wanted. (see body)
In most cases, this might be fine, but when generating
random indices between 0 and the array's capacity, you want it like
[0, capacity), and not [0, capacity]...otherwise you will be out-of-bounds.

This is what was causing memory/"double-free" errors, every few runs
of the current testing bench.

Also did work on determining sizes of the subsections each
process/thread will search in the array overall.

capacity = randrnge(ARR_RANGE_START, ARR_RANGE_END)
ARR_RANGE_START is the minimum size, ARR_RANGE_END is the maximum size.

subcap = randrnge(4, ARR_SUBCAP)
4 is the minimum subsection size, ARR_SUBCAP is the maximum subsection size.

capacity % subcap must be == 0, and subcap must be < capacity,
otherwise the values must be regenerated until this condition is met.

This ensures each process/thread searches the same quantity of values.

They have the opportunity to do an equal amount of work,
and it will also be easier for me to deal with 'fringe cases',
where the remainder of the array doesn't divide evenly with the assigned subcap...
rather, I don't have to deal with this problem at all because each capacity/subcap
pair will divide evenly anyway, based on the conditions set.

It will take a few iterations to generate the correct combination of values, but the amount of
runtime it requires to create and scramble the array, from start to finish,
will NOT be factored into the linear search time)

Will be abstracting the linear search functionality soon, and work on getting
this working with processes first.