Length of incoming buffer is calculated once, also fixed PUTMG bug
Length of incoming buffer should be calculated once; that figure can be reused as needed, since it does not change -- so why recalculate it?

Also -- let's say a user attempts to use the `put` command when a box isn't open. (`put` is designed to work when a user has opened an existing box on their client machine) -- the message will not be added to a box, but that buffer for the `put` command had not been cleared.

Now, if a client opens a user box, and uses the `put` command, but simply hits RETURN at the prompt (no data provided)...the contents of the buffer from the previous `put` invocation will be used.

This bug is now resolved with this commit, and buffers are cleared as needed.